+++ src/main/java/springdemo/databasejdbc/service/ServiceRetention.java
@@ -7,10 +7,9 @@ import java.util.List;
 public interface ServiceRetention {
 
 
-
     public RetentionModel createRetention(RetentionModel retentionModel);
-    public RetentionModel updateRetention(long id, RetentionModel retentionModel) ;
+    public RetentionModel updateRetention(int id, RetentionModel retentionModel) ;
     public List<RetentionModel> getAllRetention();
-    public RetentionModel getRetentionById(long id);
-    public void deleteRetention(Long id);
+    public RetentionModel getRetentionById(int id);
+    public void deleteRetention(int id);
 }
